+++ src/App.jsx
@@ -1,20 +1,15 @@
-import './App.css'
+import './App.css';
+import TodoList from './TodoList';
+import TodoForm from './TodoForm';
 
 function App() {
-  const todos = [
-    { id: 1, title: "Review Resources"},
-    { id: 2, title: "Take Notes"},
-    { id: 3, title: "Code Out App"},
-  ]
-
   return (
    <div>
-    <h1>My Todos</h1>
-    <ul>
-    {todos.map(todo => <li key={todo.id}>{todo.title}</li>)}
-    </ul>
+    <h1>My Todo List</h1>
+    <TodoForm />
+    <TodoList />
    </div>
-  )
-}
+  );
+};
 
-export default App
+export default App;
