+++ src/TodoList.jsx
@@ -1,0 +1,14 @@
+function TodoList() {
+    const todos = [
+        { id: 1, title: "Review Resources"},
+        { id: 2, title: "Take Notes"},
+        { id: 3, title: "Code Out App"},
+      ];
+    return (
+        <ul>
+        {todos.map(todo => <li key={todo.id}>{todo.title}</li>)}
+        </ul>
+    );
+};
+
+export default TodoList;
